# nextjs/public/openapi.yaml
openapi: 3.1.0
info:
  title: VidTempla API
  version: "1.0"
  description: |
    VidTempla is an API-first platform for managing YouTube video descriptions with templates, containers, and variables.
    The REST API lets AI agents securely manage YouTube channels without handling OAuth directly.

    ## Authentication
    All requests require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer vtk_your_key_here
    ```
    API keys have one of two permission levels:
    - **read** — can only fetch data (GET requests and POST /analytics/query)
    - **read-write** — can also create, update, and delete resources

    Generate keys from the dashboard at Settings > API Keys.

    ## Response Envelope
    Every response uses a standard envelope:
    ```json
    { "data": { ... }, "error": null, "meta": { ... } }
    ```

    ## Pagination
    All list endpoints use cursor-based pagination. Default limit is 50, max is 100.
    ```
    GET /videos?cursor=...&limit=50
    ```
  contact:
    name: VidTempla
    url: https://vidtempla.com
  license:
    name: Proprietary
servers:
  - url: https://vidtempla.com/api/v1
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key starting with vtk_. Generate from Settings > API Keys."
  schemas:
    ErrorResponse:
      type: object
      properties:
        data:
          type: "null"
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            suggestion:
              type: string
              description: Actionable suggestion for agents to self-correct
            status:
              type: integer
              description: HTTP status code
          required: [code, message, status]
      required: [data, error]
...
    SuccessMessage:
      type: object
      properties:
        data:
          type: object
          properties:
            success:
              type: boolean
              example: true
        error:
          type: "null"
...
    VideoVariable:
      type: object
      properties:
        id:
          type: string
          format: uuid
        videoId:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        variableName:
          type: string
          example: "websiteUrl"
          description: "Variable name (without {{ }} braces) as defined in the template"
        variableValue:
          type: string
          nullable: true
          example: "https://example.com"
          description: "Current value for this variable on this specific video"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        template:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
...
  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INVALID_API_KEY
              message: API key not found or has been revoked
              suggestion: Generate a new key from Settings > API Keys
              status: 401
    InsufficientPermissions:
      description: API key lacks write permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INSUFFICIENT_PERMISSIONS
              message: This API key has read-only access
              suggestion: Create a new API key with read-write permission from Settings > API Keys
              status: 403
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            data: null
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred
              suggestion: Try again later
              status: 500
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /videos/{id}/variables:
    get:
      operationId: listVideoVariables
      summary: List video variables
      description: Returns all template variables for a video, including which template each variable belongs to.
      tags: [Videos]
      x-quota-cost: 0
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: VidTempla UUID or YouTube video ID (e.g. `dQw4w9WgXcQ`)
      responses:
        "200":
          description: List of variables
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/VideoVariable"
                  error:
                    type: "null"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      operationId: updateVideoVariables
      summary: Update video variables
      description: Upserts template variables for a video and triggers a description rebuild job.
      tags: [Videos]
      x-quota-cost: 0
      x-requires-write: true
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: VidTempla UUID or YouTube video ID (e.g. `dQw4w9WgXcQ`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [variables]
              properties:
                variables:
                  type: array
                  items:
                    type: object
                    required: [templateId, name, value]
                    properties:
                      templateId:
                        type: string
                        format: uuid
                      name:
                        type: string
                        description: Variable name (without {{ }} braces)
                      value:
                        type: string
                        description: Variable value
            example:
              variables:
                - templateId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  name: "websiteUrl"
                  value: "https://example.com"
                - templateId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  name: "twitterHandle"
                  value: "@mychannel"
      responses:
        "200":
          description: Variables updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/InsufficientPermissions"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─── Templates ──────────────────────────────────────────────
...